fixed and added some padding. i changed from dvipng to imagemagick because dvipng only doesnt play nicely with variable borders and imagemagick is the best and so it does

+++ services/latex_service.rb
@@ -3,9 +3,7 @@ class LatexService
   # renders the message
   def self.render?(message, path, file)
     # stripping it so you cant just put in one letter or a string of backslashs
-    if message.strip.length == 1 || message.strip == '\\' * message.length
-      return false
-    end
+    return false if message.strip.length == 1 || message.strip == '\\' * message.length
 
     write2file(message, path, file)
 
@@ -13,15 +11,14 @@ class LatexService
     # put output into /dev/null gets rid of output
     did_comp = system("latex -interaction=nonstopmode -output-directory=#{path} #{File.join(path, file)}.tex >>/dev/null")
 
-
     # changes the .dvi to .png 
     # -q* makes it quiet
     # -D is resolution or "density" 
     # -T is image size
     # -o is output file
     if did_comp
-      system("dvipng -q* -D 300 -T tight #{File.join(path, file)}.dvi -o #{File.join(path, file)}.png")
-    else 
+      system("convert -density 300 -flatten #{File.join(path, file)}.dvi #{File.join(path, file)}.png >>/dev/null")
+    else
       return false
     end
   end
@@ -41,11 +38,9 @@ class LatexService
   def self.cleanup(path, file)
     # these are the files that are created
     file_endings = ['.aux', '.log', '.dvi', '.png', '.tex']
-    file_endings.each{ |fending|
-      if File.exist?(File.join(path, file + fending))
-        File.delete(File.join(path, file + fending))
-      end
-    }
+    file_endings.each do |fending|
+      File.delete(File.join(path, file + fending)) if File.exist?(File.join(path, file + fending))
+    end 
     #have to return nil or something will be send as a message
     nil
   end
@@ -57,10 +52,9 @@ class LatexService
     # \\ is obvious
     # " could cause an escape of the latex function
     res_chars = ['$', '\\', '"']
-    res_chars.each{ |res|
+    res_chars.each do |res|
       message = message.gsub(res, "\\#{res}")
-    }
+    end
     message
   end
-
 end
